pyproject.toml
[project]
name = "project-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "arxiv>=2.4.0",
    "faiss-cpu>=1.13.2",
    "fastapi>=0.129.0",
    "ipykernel>=7.2.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.13",
    "langchain-google-genai>=4.2.0",
    "langchain-groq>=1.1.2",
    "langchain-openai>=1.1.9",
    "langchain-tavily>=0.2.17",
    "langgraph>=1.0.8",
    "langgraph-cli[inmem]>=0.4.12",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.40.0",
    "watchdog>=6.0.0",
    "wikipedia>=1.4.0",
]
